move tests & clean up
