+++ src/lib/relay/middlewares/__tests__/cacheMiddleware-tests.ts
@@ -90,67 +90,6 @@ describe("cacheMiddleware", () => {
     })
   })
 
-  describe("a 404 response from metaphysics", () => {
-    beforeAll(() => {
-      // @ts-ignore
-      __DEV__ = false
-    })
-    afterAll(() => {
-      // @ts-ignore
-      __DEV__ = true
-    })
-    it(`will be retried if the query id was not recognized by MP`, async () => {
-      let rejected = false
-      let retried = false
-
-      const mockedErrorsNext = (req: any) => {
-        if (JSON.parse(req.fetchOpts.body).documentID) {
-          rejected = true
-          return Promise.reject(new Error("Unable to serve persisted query with ID"))
-        } else {
-          retried = true
-          return Promise.resolve({
-            status: 200,
-            json: { data: { success: true }, errors: undefined },
-          })
-        }
-      }
-      // @ts-ignore
-      await expect(cacheMiddleware()(mockedErrorsNext)(request)).resolves.toMatchObject({
-        json: { data: { success: true } },
-      })
-
-      expect(rejected).toBe(true)
-      expect(retried).toBe(true)
-    })
-
-    it(`will be not be retried if failure was something else`, async () => {
-      let rejected = false
-      let retried = false
-
-      const mockedErrorsNext = (req: any) => {
-        if (JSON.parse(req.fetchOpts.body).documentID) {
-          rejected = true
-          return Promise.reject(new Error("something unrecognized went wrong"))
-        } else {
-          retried = true
-          return Promise.resolve({
-            status: 200,
-            json: { data: { success: true }, errors: undefined },
-          })
-        }
-      }
-
-      // @ts-ignore
-      await expect(cacheMiddleware()(mockedErrorsNext)(request)).rejects.toEqual(
-        new Error("something unrecognized went wrong")
-      )
-
-      expect(rejected).toBe(true)
-      expect(retried).toBe(false)
-    })
-  })
-
   describe("a 500 response from metaphysics", () => {
     it("clears the cache and throws an error", async () => {
       const mockedErrorsNext = () => {
